04_observability_feedback_system/requirements.txt
langsmith>=0.1.75
arize-phoenix>=4.0.0
opentelemetry-sdk>=1.25.0
openai>=1.30.0
langchain>=0.2.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
pandas>=2.2.0
